move infrastructure folder to src
