+++ src/app.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { HealthModule } from './health/infrastructure/modules/health.module';
-import { UsersModule } from './users/infrastructure/modules/users.module';
+import { UsersModule } from './infrastructure/modules/users.module';
 import { ApplicationBootstrapOptions } from './users/interfaces/application-bootstrap-options.interface';
 import { CoreModule } from './users/core/core.module';
-import { UserInfrastructureModule } from './users/infrastructure/persistence/aluno-infrastructure.module';
+import { UserInfrastructureModule } from './infrastructure/persistence/aluno-infrastructure.module';
 
 @Module({
   imports: [HealthModule, CoreModule],
