+++ src/users/application/services/users.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { UpdateUserDto } from '../../http-server/dto/update-user.dto';
 import { CreateUserCommand } from '../commands/create-user.command';
 import { User } from 'src/users/domain/user';
-import { UserEntity } from 'src/users/infrastructure/persistence/typeorm/entities/user.entity';
+import { UserEntity } from 'src/infrastructure/persistence/typeorm/entities/user.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
